feat: pixel snapping for a low-DPI monitor
<https://stackoverflow.com/questions/34229483/why-is-my-svg-line-blurry-or-2px-in-height-when-i-specified-1px>

Since creating a fill-only shape in svgbob requires explicit CSS
specification, we can assume it to be an advanced feature. Therefore,
transforming a whole image for the sake of antediluvian monitors
doesn't have a serious downside for the time being. We'll revert this
with no hesitation as soon as its downside becomes apparent.