chore: bump toolchain to 1.73, invoke Cargo directly in CI (#10)
actions-rs has been deprecated. We can actually directly invoke Cargo
without preparation steps because the GitHub Actions runner image has
rustup preinstalled, and rustup automatically installs the toolchain
specified by a toolchain file.

Because the MSRV-aware resolver has not been stabilized in Cargo, and we do
not commit `Cargo.lock` (following the default behavior of an older version
of Cargo), Cargo always tries to pick the latest versions of dependencies
on every CI run even though they are incompatible with the toolchain we
specified by a toolchain file. To fix CI, this commit bumps the toolchain to
Rust 1.73.0.